Introduce `ZombieBase` entity: extend `Entity` base class, implement initialization, asset loading, animations, and orientation handling; integrate `ZombieBase` into `Game` loop and states.

+++ Core/Entities/Player.cpp
@@ -60,7 +60,7 @@ void Player::Render(SDL_Renderer* renderer)
 
 
     AnimationData& animData = animationInfo[playerCurrentTexture];
-    srcRect.w = animData.frameWidth;
+    srcRect.w = static_cast<float>(animData.frameWidth);
     srcRect.x = static_cast<float> (currentFrame) * srcRect.w;
 
     if (currentFrame >= animData.totalFrames) {
